Added Language Support

+++ src/main/java/com/example/permissionmodule/service/EmployeeService.java
@@ -3,16 +3,20 @@ package com.example.permissionmodule.service;
 import com.example.permissionmodule.dto.EmployeeDto;
 import com.example.permissionmodule.entity.Employee;
 import com.example.permissionmodule.entity.Permission;
+import com.example.permissionmodule.language.LanguageConfig;
 import com.example.permissionmodule.mapper.EmployeeMapper;
 import com.example.permissionmodule.repository.EmployeeRepository;
 import com.example.permissionmodule.repository.PermissionRepository;
 import com.example.permissionmodule.repository.PermissionRequestRepository;
+import com.example.permissionmodule.repository.PermissionResponseRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
+import java.util.Locale;
 import java.util.Optional;
 
 @Service
@@ -21,15 +25,16 @@ public class EmployeeService {
     private final EmployeeRepository employeeRepository;
     private final PermissionRepository permissionRepository;
     private final PermissionRequestRepository permissionRequestRepository;
+    private final PermissionResponseRepository permissionResponseRepository;
     private final EmployeeMapper employeeMapper;
 
-    public EmployeeService(EmployeeRepository employeeRepository, PermissionRepository permissionRepository, PermissionRequestRepository permissionRequestRepository, EmployeeMapper employeeMapper) {
+    public EmployeeService(EmployeeRepository employeeRepository, PermissionRepository permissionRepository, PermissionRequestRepository permissionRequestRepository, PermissionResponseRepository permissionResponseRepository, EmployeeMapper employeeMapper) {
         this.employeeRepository = employeeRepository;
         this.permissionRepository = permissionRepository;
         this.permissionRequestRepository = permissionRequestRepository;
+        this.permissionResponseRepository = permissionResponseRepository;
         this.employeeMapper = employeeMapper;
     }
-
 
     public List<Employee> findAll() {
         return employeeRepository.findAll();
@@ -47,7 +52,7 @@ public class EmployeeService {
         return employeeRepository.findById(employeeId);
     }
 
-    public Employee updateEmployee(Long employeeId, Employee newEmployee) {
+    public Employee updateEmployee(Long employeeId, Employee newEmployee) throws NotFoundException {
         Optional<Employee> employee = employeeRepository.findById(employeeId);
         if (employee.isPresent()) {
             Employee foundEmployee = employee.get();
@@ -57,18 +62,21 @@ public class EmployeeService {
             employeeRepository.save(foundEmployee);
             return foundEmployee;
         } else
-            return null;
+            throw new NotFoundException(LanguageConfig.getErrorMessage("employee.not.found",Locale.US.getLanguage()));
     }
 
     @Transactional
     public void deleteById(Long employeeId) {
 
-        permissionRequestRepository.deleteByEmployeeId(employeeId);
-
-        permissionRepository.deleteByEmployeeId(employeeId);
-
-        employeeRepository.deleteById(employeeId);
-
+        Optional<Employee> employee = employeeRepository.findById(employeeId);
+        if (employee.isPresent()) {
+            permissionResponseRepository.deleteById(employeeId);
+            permissionRequestRepository.deleteByEmployeeId(employeeId);
+            permissionRepository.deleteByEmployeeId(employeeId);
+            employeeRepository.deleteById(employeeId);
+        }
+        else
+            throw new NotFoundException(LanguageConfig.getErrorMessage("employee.not.found",Locale.US.getLanguage()));
     }
 
     public long getWorkTime(Employee employee){
